refactor(workspace): extract worktree lifecycle into actions/worktree.rs
actions/project.rs (1222 lines) mixed project CRUD/ordering/sizing with
the entire worktree lifecycle. Move the worktree methods (create/register/
discover/remove/reorder + their helpers and tests) into a new
actions/worktree.rs, matching how the actions module already splits by
domain. project.rs drops to 772 lines; everything stays methods on
Workspace so no re-exports change.

Also routes the repeated self.data.projects.iter_mut().find(|p| p.id ==
...) sites through the existing project_mut(id) helper. Pure refactor:
176 tests pass identically, no behavior or API change.
