+++ crates/okena-workspace/src/actions/mod.rs
@@ -5,12 +5,14 @@
 //! - `layout`: Split, tabs, and close operations
 //! - `project`: Project CRUD and properties
 //! - `terminal`: Terminal-specific actions
+//! - `worktree`: Worktree lifecycle (create, register, discover, remove)
 
 pub mod focus;
 pub mod folder;
 pub mod layout;
 pub mod project;
 pub mod terminal;
+pub mod worktree;
 
 // All impl blocks are on Workspace, so no re-exports needed.
 // The methods are available directly on the Workspace type.
